requirements.txt
werkzeug==3.0.3
Flask==3.0.3
flask_sqlalchemy
python-dotenv==1.0.1
flask_login
flask_wtf
WTForms==3.1.2
email_validator==2.1.1
pyotp
flask-qrcode
MarkupSafe==2.1.5
cryptography==42.0.7
bcrypt~=4.1.3
rsa~=4.9
Flask-Talisman
requests~=2.31.0
bs4~=0.0.2
flask_paginate
pydub
pyzbar~=0.1.9
opencv-python~=4.9.0.80
beautifulsoup4~=4.12.3
soupsieve~=2.5
cffi~=1.16.0
pillow~=10.3.0
pip~=21.3.1
wheel~=0.37.1
pyasn1~=0.6.0
click~=8.1.7
config~=0.5.1
Jinja2~=3.1.2
blinker~=1.7.0
itsdangerous~=2.1.2
qrcode~=7.4.2
colorama~=0.4.6
setuptools~=60.2.0
idna~=3.6
dnspython~=2.4.2
SQLAlchemy~=2.0.23
selenium~=4.21.0